Internal functions also used in Voyager

Not meant for the user, but exporting to be used internally in Voyager. But one day I may clean these up and remove the internal note for people building on top of SFE.

Usage

.value2df(value, use_geometry, feature = NULL)

.check_features(x, features, colGeometryName = NULL, swap_rownames = NULL)

.warn_symbol_duplicate(x, symbols, swap_rownames = "symbol")

.symbol2id(x, features, swap_rownames)

.check_sample_id(x, sample_id, one = TRUE, mustWork = TRUE)

.rm_empty_geometries(g, MARGIN)

.check_rg(type, x, sample_id)

.ext_(x)

Value

Internal