Update a SpatialFeatureExperiment object

Update a SpatialFeatureExperiment to the latest version of object structure. This is usually called by internal functions.

Usage

# S4 method for SpatialFeatureExperiment
updateObject(object, ..., verbose = FALSE)

SFEVersion(object)

Arguments

object

An old SpatialFeatureExperiment object.

...

Additional arguments that are ignored.

verbose

Logical scalar indicating whether a message should be emitted as the object is updated.

Value

An updated version of object.

Details

Version 1.1.4 adds package version to the SFE object. We are considering an overhaul of the spatialGraphs slot in a future version using the sfdep package to decouple the adjacency graph from the edge weights.

See also

Examples

library(SFEData)
sfe <- McKellarMuscleData("small")
#> see ?SFEData and browseVignettes('SFEData') for documentation
#> loading from cache
# First version of SFE object doesn't log SFE package version, so should be NULL
SFEVersion(sfe)
#> [1] ‘1.2.2’
sfe <- updateObject(sfe)
# See current version
SFEVersion(sfe)
#> [1] ‘1.2.2’