Get the image from *Image class

In SFE, S4 classes inheriting from VirtualSpatialImage have been implemented to make these image classes compatible with SpatialExperiment. The imgRaster methods in SFE are meant to extract the original image from the *Image classes, such as SpatRaster from SpatRasterImage, and Image from ExtImage and BioFormatsImage. For BioFormatsImage, the image of the specified resolution will be read into memory as AnnotatedImage, which inherits from EBImage::Image.

Arguments

x

An object of class *Image as implemented in this package.

resolution

Resolution to read in from OME-TIFF, defaults to 4, which is a medium resolution in Xenium.

Value

SpatRaster from SpatRasterImage, and Image from ExtImage and BioFormatsImage. For BioFormatsImage, the image of the specified resolution will be read into memory as AnnotatedImage and ExtImage, which both inherit from EBImage::Image.

See also